backup_tests/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_community
langchain-core
langchain-groq
groq
python-dotenv
gtts
SpeechRecognition
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart